Removed Encoding Logic From App.js
