+++ src/App.js
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { createFFmpeg, fetchFile } from '@ffmpeg/ffmpeg';
+import { createFFmpeg } from '@ffmpeg/ffmpeg';
 import styles from './App.module.scss';
 
 import MyDropzone from './Components/MyDropzone/MyDropzone';
 import Button from './Components/Button/Button';
 import Selector from './Components/Selector/Selector';
 import ProgressBar from './Components/ProgressBar/ProgressBar';
+import EncodeToSize from './Util/EncodeToSize';
 
 const ffmpeg = createFFmpeg({ log: true });
 
@@ -33,56 +34,13 @@ function App() {
     setBufferSize(bitrate * 2);
   }, [targetSize, bitrate, duration, encoded]);
 
-  const encodeToSize = async () => {
-    // Write the file to memory
-    ffmpeg.FS('writeFile', 'inputFile.mp4', await fetchFile(video));
-
-    // Run the FFMpeg command
-    if (targetSize >= 50) {
-      await ffmpeg.run(
-        '-i', 'inputFile.mp4',
-        '-i', 'inputFile.mp4',
-        '-vf', 'scale=1920:1080',
-        '-c:v', 'libx264',
-        '-preset', 'superfast',
-        '-b:v', `${bitrate}k`,
-        '-minrate', `${bitrate}k`,
-        '-maxrate', `${bitrate}k`,
-        '-bufsize', `${bufferSize}k`,
-        '-ac', '2',
-        '-c:a', 'aac',
-        '-b:a', '128k',
-        'encoded.mp4',
-      );
-    } else {
-      await ffmpeg.run(
-        '-i', 'inputFile.mp4',
-        '-vf', 'scale=1280:720',
-        '-r', '30',
-        '-c:v', 'libx264',
-        '-preset', 'superfast',
-        '-b:v', `${bitrate}k`,
-        '-minrate', `${bitrate}k`,
-        '-maxrate', `${bitrate}k`,
-        '-bufsize', `${bufferSize}k`,
-        '-ac', '2',
-        '-c:a', 'aac',
-        '-b:a', '128k',
-        'encoded.mp4',
-      );
-    }
-
-    // Read the result
-    const data = ffmpeg.FS('readFile', 'encoded.mp4');
-
-    // Create a URL
-    const url = URL.createObjectURL(new Blob([data.buffer], { type: 'video/mp4' }));
-    setEncoded(url);
-  };
-
   ffmpeg.setProgress(({ ratio }) => {
     setProgress(ratio * 100);
   });
+
+  const startEncoding = () => {
+    setEncoded(EncodeToSize(video, targetSize, bitrate, bufferSize));
+  };
 
   const renderSwitch = () => {
     if (encoded) {
@@ -114,7 +72,7 @@ function App() {
         </div>
         <Selector progress={progress} tSize={targetSize} setTSize={setTargetSize} />
 
-        <Button sEncode={encodeToSize} vStatus={video}>
+        <Button sEncode={startEncoding} vStatus={video}>
           {!video ? 'Upload A Video' : 'Convert'}
         </Button>
       </>
